refactor(tasks): mueve CopyTextButton a shared/ui/atoms
